remove old sample
