+++ example/App.hs
@@ -3,6 +3,8 @@
 {-# LANGUAGE OverloadedStrings         #-}
 {-# LANGUAGE RankNTypes                #-}
 {-# LANGUAGE RecordWildCards           #-}
+{-# LANGUAGE TypeFamilies              #-}
+{-# LANGUAGE GADTs              #-}
 
 module App (app, waiApp) where
 
@@ -23,12 +25,17 @@ import           Network.Wai.Middleware.Static
 import           Prelude
 import           Web.Scotty
 import           Web.Scotty.Internal.Types
+import qualified Data.HashMap.Strict     as Map
 
 import           IDP
 import           Session
 import           Types
 import           Utils
 import           Views
+
+import qualified IDP.Github           as IGithub
+import qualified IDP.Facebook         as IFacebook
+
 
 ------------------------------
 -- App
@@ -45,8 +52,8 @@ app = putStrLn ("Starting Server. http://localhost:" ++ show myServerPort)
 -- TODO: how to add either Monad or a middleware to do session?
 waiApp :: IO WAI.Application
 waiApp = do
-  cache <- initKeyCache
-  idps cache
+  cache <- initCacheStore
+  initIdps cache
   scottyApp $ do
     middleware $ staticPolicy (addBase "example/assets")
     defaultHandler globalErrorHandler
@@ -54,26 +61,12 @@ waiApp = do
     get "/oauth2/callback" $ callbackH cache
     get "/logout" $ logoutH cache
 
+debug :: Bool
+debug = True
 
 --------------------------------------------------
 -- * Handlers
 --------------------------------------------------
-
-idps :: KeyCache -> IO ()
-idps c =
-  -- TODO: leverage generic ?
-  mapM_ (\idp -> insertKeys c (idpName idp) idp)
-  (fmap mkIDPData [ Okta
-                  , Github
-                  , Google
-                  , Douban
-                  , Dropbox
-                  , Facebook
-                  , Fitbit
-                  , Weibo
-                  , StackExchange
-                  ]
-  )
 
 redirectToHomeM :: ActionM ()
 redirectToHomeM = redirect "/"
@@ -84,62 +77,62 @@ errorM = throwError . ActionError
 globalErrorHandler :: Text -> ActionM ()
 globalErrorHandler t = status status401 >> html t
 
-logoutH :: KeyCache -> ActionM ()
+logoutH :: CacheStore -> ActionM ()
 logoutH c = do
   pas <- params
   let idpP = paramValue "idp" pas
   when (null idpP) redirectToHomeM
-  let maybeIdp = idpFromText (head idpP)
+  let maybeIdp = parseIDP (head idpP)
   case maybeIdp of
-    Just idpInput -> liftIO (removeKey c idpInput) >> redirectToHomeM
-    Nothing       -> errorM ("logout: unknown IDP " `TL.append` head idpP)
+    Right (IDPApp idp) -> liftIO (removeKey c (idpLabel idp)) >> redirectToHomeM
+    Left e       -> errorM ("logout: unknown IDP " `TL.append` e)
 
-indexH :: KeyCache -> ActionM ()
+indexH :: CacheStore -> ActionM ()
 indexH c = do
   is <- liftIO (allValues c)
   overviewTpl is
 
-callbackH :: KeyCache -> ActionM ()
+callbackH :: CacheStore -> ActionM ()
 callbackH c = do
   pas <- params
   let codeP = paramValue "code" pas
   let stateP = paramValue "state" pas
   when (null codeP) (errorM "callbackH: no code from callback request")
   when (null stateP) (errorM "callbackH: no state from callback request")
-  let maybeIdpName = idpFromText $ TL.takeWhile (/= '.') (head stateP)
+  let maybeIdpName = parseIDP (TL.takeWhile (/= '.') (head stateP))
   -- TODO: looks like `state` shall be passed when fetching access token
   --       turns out no IDP enforce this yet
   case maybeIdpName of
-    Just idpN -> fetchTokenAndUser (head codeP) c idpN
-    Nothing   -> errorM ("callbackH: cannot find IDP name from text " `TL.append` head stateP)
+    Right (IDPApp idp) -> fetchTokenAndUser c (head codeP) idp
+    Left e   -> errorM ("callbackH: cannot find IDP name from text " `TL.append` head stateP)
 
-fetchTokenAndUser :: TL.Text           -- ^ code
-                  -> KeyCache
-                  -> IDP
+fetchTokenAndUser :: (HasTokenReq a, HasUserReq a, HasLabel a) => CacheStore
+                  -> TL.Text           -- ^ code
+                  -> a
                   -> ActionM ()
-fetchTokenAndUser code store idpInput = do
-  mayBeIdp <- liftIO $ lookupKey store idpInput
+fetchTokenAndUser store code idpInput = do
+  mayBeIdp <- liftIO $ lookupKey store (idpLabel idpInput)
   when (isNothing mayBeIdp) (errorM "fetchTokenAndUser: cannot find idp data from cache")
   let idpD = fromJust mayBeIdp
   result <- liftIO $ do
     mgr <- newManager tlsManagerSettings
-    token <- tryFetchAT idpD mgr (ExchangeToken $ TL.toStrict code)
-    --print token
+    token <- tryFetchAT idpInput mgr (ExchangeToken $ TL.toStrict code)
+    when debug (print token)
     case token of
-      Right at -> tryFetchUser idpD mgr (accessToken at)
+      Right at -> tryFetchUser idpInput mgr (accessToken at)
       Left e   -> return (Left $ TL.pack $ "cannot fetch asses token. error detail: " ++ show e)
   case result of
     Right lUser -> do
       let newIdp = idpD {loginUser = Just lUser }
-      liftIO $ insertKeys store idpInput newIdp
+      liftIO $ insertIDPData store newIdp
       redirectToHomeM
     Left err -> errorM ("fetchTokenAndUser: " `TL.append` err)
 
 -- * Fetch UserInfo
 --
-tryFetchUser :: IDPData -> Manager -> AccessToken -> IO (Either Text LoginUser)
-tryFetchUser IDPData {..} mgr token = do
-  re <- getUserInfo mgr token
+tryFetchUser :: (HasUserReq a) => a -> Manager -> AccessToken -> IO (Either Text LoginUser)
+tryFetchUser idp mgr token = do
+  re <- userReq idp mgr token
   return (first displayOAuth2Error re)
 
 displayOAuth2Error :: OAuth2Error Errors -> Text
@@ -147,8 +140,8 @@ displayOAuth2Error = TL.pack . show
 
 -- * Fetch Access Token
 --
-tryFetchAT :: IDPData
+tryFetchAT :: (HasTokenReq a) => a
   -> Manager
   -> ExchangeToken
   -> IO (OAuth2Result TR.Errors OAuth2Token)
-tryFetchAT IDPData {..} mgr = getAccessToken mgr oauth2Key
+tryFetchAT idp mgr code = tokenReq idp mgr code
