+++ example/IDP/Douban.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE QuasiQuotes   #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module IDP.Douban where
 import           Data.Aeson
@@ -13,6 +14,29 @@ import qualified Network.OAuth.OAuth2.TokenRequest as TR
 import           Types
 import           URI.ByteString
 import           URI.ByteString.QQ
+import           Data.Hashable
+import Keys
+import Utils
+
+data Douban = Douban deriving (Show, Generic)
+
+instance Hashable Douban
+
+instance IDP Douban
+
+instance HasLabel Douban
+
+instance HasTokenReq Douban where
+  tokenReq _ mgr code = fetchAccessToken2 mgr doubanKey code
+
+instance HasUserReq Douban where
+  userReq _ mgr at = do
+    re <- authGetJSON mgr at userInfoUri
+    return (second toLoginUser re)
+
+instance HasAuthUri Douban where
+  authUri _ = createCodeUri doubanKey [ ("state", "Douban.test-state-123")
+                                        ] 
 
 data DoubanUser = DoubanUser { name :: Text
                              , uid  :: Text
