+++ example/IDP/Dropbox.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE QuasiQuotes   #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module IDP.Dropbox where
 import           Data.Aeson
@@ -13,6 +14,29 @@ import qualified Network.OAuth.OAuth2.TokenRequest as TR
 import           Types
 import           URI.ByteString
 import           URI.ByteString.QQ
+import           Data.Hashable
+import Keys
+import Utils
+
+data Dropbox = Dropbox deriving (Show, Generic)
+
+instance Hashable Dropbox
+
+instance IDP Dropbox
+
+instance HasLabel Dropbox
+
+instance HasTokenReq Dropbox where
+  tokenReq _ mgr code = fetchAccessToken mgr dropboxKey code
+
+instance HasUserReq Dropbox where
+  userReq _ mgr at = do
+    re <- parseResponseJSON <$> authPostBS3 mgr at userInfoUri
+    return (second toLoginUser re)
+
+instance HasAuthUri Dropbox where
+  authUri _ = createCodeUri dropboxKey [ ("state", "Dropbox.test-state-123")
+                                        ] 
 
 newtype DropboxName = DropboxName { displayName :: Text }
                  deriving (Show, Generic)
@@ -23,6 +47,7 @@ data DropboxUser = DropboxUser { email :: Text
 
 instance FromJSON DropboxName where
     parseJSON = genericParseJSON defaultOptions { fieldLabelModifier = camelTo2 '_' }
+
 instance FromJSON DropboxUser where
     parseJSON = genericParseJSON defaultOptions { fieldLabelModifier = camelTo2 '_' }
