+++ example/IDP/Facebook.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE QuasiQuotes   #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module IDP.Facebook where
 import           Data.Aeson
@@ -13,6 +14,30 @@ import qualified Network.OAuth.OAuth2.TokenRequest as TR
 import           Types
 import           URI.ByteString
 import           URI.ByteString.QQ
+import           Data.Hashable
+import Keys
+import Utils
+
+data Facebook = Facebook deriving (Show, Generic)
+
+instance Hashable Facebook
+
+instance IDP Facebook
+
+instance HasLabel Facebook
+
+instance HasTokenReq Facebook where
+  tokenReq _ mgr code = fetchAccessToken2 mgr facebookKey code
+
+instance HasUserReq Facebook where
+  userReq _ mgr at = do
+    re <- authGetJSON mgr at userInfoUri
+    return (second toLoginUser re)
+
+instance HasAuthUri Facebook where
+  authUri _ = createCodeUri facebookKey [ ("state", "Facebook.test-state-123")
+                                        , ("scope", "user_about_me,email")
+                                        ] 
 
 data FacebookUser = FacebookUser { id    :: Text
                                  , name  :: Text
