+++ example/IDP/Okta.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE QuasiQuotes   #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module IDP.Okta where
 import           Data.Aeson
@@ -13,6 +14,30 @@ import qualified Network.OAuth.OAuth2.TokenRequest as TR
 import           Types
 import           URI.ByteString
 import           URI.ByteString.QQ
+import           Data.Hashable
+import Keys
+import Utils
+
+data Okta = Okta deriving (Show, Generic)
+
+instance Hashable Okta
+
+instance IDP Okta
+
+instance HasLabel Okta
+
+instance HasTokenReq Okta where
+  tokenReq _ mgr code = fetchAccessToken mgr oktaKey code
+
+instance HasUserReq Okta where
+  userReq _ mgr at = do
+    re <- authGetJSON mgr at userInfoUri
+    return (second toLoginUser re)
+
+instance HasAuthUri Okta where
+  authUri _ = createCodeUri oktaKey [ ("state", "Okta.test-state-123")
+                                        , ("scope", "openid profile")
+                                        ] 
 
 data OktaUser = OktaUser { name              :: Text
                          , preferredUsername :: Text
