+++ example/Types.hs
@@ -20,7 +20,33 @@ import qualified Network.OAuth.OAuth2.TokenRequest as TR
 import           Text.Mustache
 import qualified Text.Mustache                     as M
 import           URI.ByteString
+import           Data.ByteString      (ByteString)
+import qualified Data.Text.Encoding   as TE
+import qualified Data.HashMap.Strict     as Map
+import           Control.Concurrent.MVar
 
+type IDPLabel = Text
+
+-- TODO: how to make following type work??
+-- type CacheStore = forall a. IDP a => MVar (Map.HashMap a IDPData)
+type CacheStore = MVar (Map.HashMap IDPLabel IDPData)
+
+-- * type class for defining a IDP
+--
+class (Hashable a, Show a) => IDP a
+
+class (IDP a) => HasLabel a where
+  idpLabel :: a -> IDPLabel
+  idpLabel = TL.pack . show
+
+class (IDP a) => HasAuthUri a where
+  authUri :: a -> Text
+    
+class (IDP a) => HasTokenReq a where
+  tokenReq :: a -> Manager -> ExchangeToken -> IO (OAuth2Result TR.Errors OAuth2Token)
+
+class (IDP a) => HasUserReq a where
+  userReq :: FromJSON b => a -> Manager -> AccessToken -> IO (OAuth2Result b LoginUser)
 
 -- dummy oauth2 request error
 --
@@ -31,33 +57,6 @@ data Errors =
 instance FromJSON Errors where
   parseJSON = genericParseJSON defaultOptions { constructorTagModifier = camelTo2 '_', allNullaryToStringTag = True }
 
-data IDP =
-    Douban
-  | Dropbox
-  | Facebook
-  | Fitbit
-  | Github
-  | Google
-  | Okta
-  | StackExchange
-  | Weibo
-  deriving (Show, Eq, Generic)
-
-instance Hashable IDP
-
-idpFromText :: Text -> Maybe IDP
-idpFromText ift = case TL.unpack $ TL.toLower ift of
-  "okta"          -> Just Okta
-  "github"        -> Just Github
-  "google"        -> Just Google
-  "douban"        -> Just Douban
-  "dropbox"       -> Just Dropbox
-  "facebook"      -> Just Facebook
-  "fitbit"        -> Just Fitbit
-  "weibo"         -> Just Weibo
-  "stackexchange" -> Just StackExchange
-  _               -> Nothing
-
 newtype LoginUser =
   LoginUser { loginUserName :: Text
             } deriving (Eq, Show)
@@ -65,18 +64,18 @@ newtype LoginUser =
 data IDPData =
   IDPData { codeFlowUri :: Text
           , loginUser   :: Maybe LoginUser
-          , idpName     :: IDP
-          , oauth2Key   :: OAuth2
-          , getAccessToken :: Manager -> OAuth2 -> ExchangeToken -> IO (OAuth2Result TR.Errors OAuth2Token)
-          , getUserInfo :: Manager -> AccessToken -> IO (OAuth2Result Errors LoginUser)
+          , idpDisplayLabel     :: IDPLabel
           }
 
 -- simplify use case to only allow one idp instance for now.
 instance Eq IDPData where
-  a == b = idpName a == idpName b
+  a == b = idpDisplayLabel a == idpDisplayLabel b
+
+instance Ord IDPData where
+  a `compare` b = idpDisplayLabel a `compare` idpDisplayLabel b
 
 newtype TemplateData = TemplateData { idpData :: [IDPData]
-                                 } deriving (Eq)
+                                    } deriving (Eq)
 
 -- * Mustache instances
 instance ToMustache IDPData where
@@ -84,7 +83,7 @@ instance ToMustache IDPData where
     [ "codeFlowUri" ~> codeFlowUri t'
     , "isLogin" ~> isJust (loginUser t')
     , "user" ~> loginUser t'
-    , "name" ~> show (idpName t')
+    , "name" ~> TL.unpack (idpDisplayLabel t')
     ]
 
 instance ToMustache LoginUser where
