--- example2/IDP/Douban.hs
@@ -1,63 +1,0 @@
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE QuasiQuotes   #-}
-{-# LANGUAGE OverloadedStrings #-}
-
-module IDP.Douban where
-import           Data.Aeson
-import           Data.Aeson.Types
-import           Data.Bifunctor
-import           Data.Text.Lazy                    (Text)
-import           GHC.Generics
-import           Network.HTTP.Conduit
-import           Network.OAuth.OAuth2
-import qualified Network.OAuth.OAuth2.TokenRequest as TR
-import           Types
-import           URI.ByteString
-import           URI.ByteString.QQ
-import           Data.Hashable
-import Keys
-import Utils
-
-data Douban = Douban deriving (Show, Generic)
-
-instance Hashable Douban
-
-instance IDP Douban
-
-instance HasLabel Douban
-
-instance HasTokenReq Douban where
-  tokenReq _ mgr code = fetchAccessToken2 mgr doubanKey code
-
-instance HasUserReq Douban where
-  userReq _ mgr at = do
-    re <- authGetJSON mgr at userInfoUri
-    return (second toLoginUser re)
-
-instance HasAuthUri Douban where
-  authUri _ = createCodeUri doubanKey [ ("state", "Douban.test-state-123")
-                                        ] 
-
-data DoubanUser = DoubanUser { name :: Text
-                             , uid  :: Text
-                             } deriving (Show, Generic)
-
-instance FromJSON DoubanUser where
-    parseJSON = genericParseJSON defaultOptions { fieldLabelModifier = camelTo2 '_' }
-
-userInfoUri :: URI
-userInfoUri = [uri|https://api.douban.com/v2/user/~me|]
-
-toLoginUser :: DoubanUser -> LoginUser
-toLoginUser ouser = LoginUser { loginUserName = name ouser }
-
-getUserInfo :: FromJSON a => Manager -> AccessToken -> IO (OAuth2Result a LoginUser)
-getUserInfo mgr at = do
-  re <- authGetJSON mgr at userInfoUri
-  return (second toLoginUser re)
-
-getAccessToken :: Manager
-               -> OAuth2
-               -> ExchangeToken
-               -> IO (OAuth2Result TR.Errors OAuth2Token)
-getAccessToken = fetchAccessToken2
