--- example2/Utils.hs
@@ -1,36 +1,0 @@
-module Utils where
-
-import qualified Data.Aeson                as Aeson
-import           Data.ByteString           (ByteString)
-import qualified Data.Text.Encoding        as T
-import           Data.Text.Lazy            (Text)
-import qualified Data.Text.Lazy            as TL
-import           Web.Scotty.Internal.Types
-import           Network.OAuth.OAuth2
-import qualified Data.Text.Lazy                    as TL
-import qualified Data.Text.Encoding   as TE
-import           URI.ByteString
-
-tlToBS :: TL.Text -> ByteString
-tlToBS = T.encodeUtf8 . TL.toStrict
-
-paramValue :: Text -> [Param] -> [Text]
-paramValue key = fmap snd . filter (hasParam key)
-
-hasParam :: Text -> Param -> Bool
-hasParam t = (== t) . fst
-
-parseValue :: Aeson.FromJSON a => Maybe Aeson.Value -> Maybe a
-parseValue Nothing = Nothing
-parseValue (Just a) = case Aeson.fromJSON a of
-  Aeson.Error _   -> Nothing
-  Aeson.Success b -> Just b
-
-createCodeUri :: OAuth2
-  -> [(ByteString, ByteString)]
-  -> Text
-createCodeUri key params = TL.fromStrict $ TE.decodeUtf8 $ serializeURIRef'
-  $ appendQueryParams params
-  $ authorizationUrl key
-
-
