--- example2/main.hs
@@ -1,6 +1,0 @@
-module Main where
-
-import           App (app)
-
-main :: IO ()
-main = app
