--- example2/templates/index.mustache
@@ -1,30 +1,0 @@
-<html>
-    <head>
-        <link href="/main.css" rel="stylesheet"/>
-    </head>
-    <body>
-        <h1>Hello OAuth2</h1>
-        {{#idps}}
-        <section class="login-with">
-
-            {{^isLogin}}
-            <a href="{{codeFlowUri}}">Login {{name}}</a>
-            {{/isLogin}}
-
-            {{#isLogin}}
-            <h2>Welcome to {{name}}</h2>
-            {{#user}}
-            <div class="result">Hello, {{name}}</div>
-            {{/user}}
-            <a href="/logout?idp={{name}}">Logout</a>
-            {{/isLogin}}
-
-        </section>
-        {{/idps}}
-        <h2>Notes</h2>
-        <ol>
-            <li>for StackExchange, the callback domain is localhost, have manually add port 9988.</li>
-        </ol>
-        </p>
-    </body>
-</html>
