+++ hoauth2.cabal
@@ -114,7 +114,7 @@ Executable demo-server
                        Types
                        Utils
                        Views
-  hs-source-dirs:      example2
+  hs-source-dirs:      example
   default-language:    Haskell2010
   build-depends:       base              >= 4.5    && < 5,
                        text              >= 0.11   && < 1.3,
